More cleanup and Refactoring
