+++ src/main/java/io/github/thebusybiscuit/slimefun4/core/attributes/Soulbound.java
@@ -4,16 +4,12 @@ import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.SlimefunItem;
 
 /**
  * This Interface, when attached to a class that inherits from {@link SlimefunItem}, marks
- * the Item as radioactive.
- * Carrying such an item will give the wearer the radiation effect.
- * 
- * You can specify a level of {@link Radioactivity} for the severity of the effect.
+ * the Item as soulbound.
+ * This Item will then not be dropped upon death.
  * 
  * @author TheBusyBiscuit
  *
  */
-public interface Radioactive {
+public interface Soulbound {
 	
-	Radioactivity getRadioactivity();
-
 }
