+++ src/main/java/me/mrCookieSlime/Slimefun/Objects/SlimefunItem/JetBoots.java
@@ -6,12 +6,12 @@ import me.mrCookieSlime.Slimefun.Lists.Categories;
 import me.mrCookieSlime.Slimefun.Lists.RecipeType;
 import me.mrCookieSlime.Slimefun.api.SlimefunItemStack;
 
-public class JetBoots extends DamagableChargableItem {
+public class JetBoots extends ChargableItem {
 	
 	private final double speed;
 
 	public JetBoots(SlimefunItemStack item, ItemStack[] recipe, double speed) {
-		super(Categories.TECH, item, RecipeType.ENHANCED_CRAFTING_TABLE, recipe, "Jet Boots");
+		super(Categories.TECH, item, RecipeType.ENHANCED_CRAFTING_TABLE, recipe);
 		this.speed = speed;
 	}
