+++ src/main/java/me/mrCookieSlime/Slimefun/Objects/SlimefunItem/Jetpack.java
@@ -6,12 +6,12 @@ import me.mrCookieSlime.Slimefun.Lists.Categories;
 import me.mrCookieSlime.Slimefun.Lists.RecipeType;
 import me.mrCookieSlime.Slimefun.api.SlimefunItemStack;
 
-public class Jetpack extends DamagableChargableItem {
+public class Jetpack extends ChargableItem {
 	
 	private final double thrust;
 
 	public Jetpack(SlimefunItemStack item, ItemStack[] recipe, double thrust) {
-		super(Categories.TECH, item, RecipeType.ENHANCED_CRAFTING_TABLE, recipe, "Jetpack");
+		super(Categories.TECH, item, RecipeType.ENHANCED_CRAFTING_TABLE, recipe);
 		this.thrust = thrust;
 	}
