+++ src/main/java/me/mrCookieSlime/Slimefun/Objects/SlimefunItem/Juice.java
@@ -6,12 +6,14 @@ import me.mrCookieSlime.Slimefun.Lists.RecipeType;
 import me.mrCookieSlime.Slimefun.Objects.Category;
 import me.mrCookieSlime.Slimefun.api.SlimefunItemStack;
 
+/**
+ * This class represents a {@link SlimefunItem} that can be stored inside
+ * of a Cooler.
+ * 
+ * @author TheBusyBiscuit
+ *
+ */
 public class Juice extends SlimefunItem {
-
-	@Deprecated
-	public Juice(Category category, ItemStack item, String id, RecipeType recipeType, ItemStack[] recipe) {
-		super(category, item, id, recipeType, recipe);
-	}
 
 	public Juice(Category category, SlimefunItemStack item, RecipeType recipeType, ItemStack[] recipe) {
 		super(category, item, recipeType, recipe);
