+++ src/main/java/me/mrCookieSlime/Slimefun/Objects/SlimefunItem/SlimefunMachine.java
@@ -21,6 +21,14 @@ public class SlimefunMachine extends SlimefunItem implements RecipeDisplayItem {
 	protected final List<ItemStack> shownRecipes;
 	protected final MultiBlock multiblock;
 
+	public SlimefunMachine(Category category, SlimefunItemStack item, ItemStack[] recipe, ItemStack[] machineRecipes, BlockFace trigger) {
+		super(category, item, RecipeType.MULTIBLOCK, recipe);
+		this.recipes = new ArrayList<>();
+		this.shownRecipes = new ArrayList<>();
+		this.shownRecipes.addAll(Arrays.asList(machineRecipes));
+		this.multiblock = new MultiBlock(this, convertItemStacksToMaterial(recipe), trigger);
+	}
+
 	@Deprecated
 	public SlimefunMachine(Category category, ItemStack item, String id, ItemStack[] recipe, ItemStack[] machineRecipes, BlockFace trigger) {
 		super(category, item, id, RecipeType.MULTIBLOCK, recipe);
@@ -30,25 +38,8 @@ public class SlimefunMachine extends SlimefunItem implements RecipeDisplayItem {
 		this.multiblock = new MultiBlock(this, convertItemStacksToMaterial(recipe), trigger);
 	}
 
-	public SlimefunMachine(Category category, SlimefunItemStack item, ItemStack[] recipe, ItemStack[] machineRecipes, BlockFace trigger) {
-		super(category, item, RecipeType.MULTIBLOCK, recipe);
-		this.recipes = new ArrayList<>();
-		this.shownRecipes = new ArrayList<>();
-		this.shownRecipes.addAll(Arrays.asList(machineRecipes));
-		this.multiblock = new MultiBlock(this, convertItemStacksToMaterial(recipe), trigger);
-	}
-
 	public SlimefunMachine(Category category, SlimefunItemStack item, ItemStack[] recipe, ItemStack[] machineRecipes, BlockFace trigger, String[] keys, Object[] values) {
 		super(category, item, RecipeType.MULTIBLOCK, recipe, keys, values);
-		this.recipes = new ArrayList<>();
-		this.shownRecipes = new ArrayList<>();
-		this.shownRecipes.addAll(Arrays.asList(machineRecipes));
-		this.multiblock = new MultiBlock(this, convertItemStacksToMaterial(recipe), trigger);
-	}
-
-	@Deprecated
-	public SlimefunMachine(Category category, ItemStack item, String id, ItemStack[] recipe, ItemStack[] machineRecipes, BlockFace trigger, String[] keys, Object[] values) {
-		super(category, item, id, RecipeType.MULTIBLOCK, recipe, keys, values);
 		this.recipes = new ArrayList<>();
 		this.shownRecipes = new ArrayList<>();
 		this.shownRecipes.addAll(Arrays.asList(machineRecipes));
