+++ src/main/java/me/mrCookieSlime/Slimefun/Objects/SlimefunItem/SoulboundItem.java
@@ -2,10 +2,10 @@ package me.mrCookieSlime.Slimefun.Objects.SlimefunItem;
 
 import org.bukkit.inventory.ItemStack;
 
+import io.github.thebusybiscuit.slimefun4.core.attributes.Soulbound;
 import me.mrCookieSlime.Slimefun.Lists.RecipeType;
 import me.mrCookieSlime.Slimefun.Objects.Category;
 import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.interfaces.NotPlaceable;
-import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.interfaces.Soulbound;
 import me.mrCookieSlime.Slimefun.api.SlimefunItemStack;
 
 /**
