--- src/main/java/me/mrCookieSlime/Slimefun/Objects/SlimefunItem/interfaces/Soulbound.java
@@ -1,5 +1,0 @@
-package me.mrCookieSlime.Slimefun.Objects.SlimefunItem.interfaces;
-
-public interface Soulbound {
-	
-}
