+++ src/main/java/me/mrCookieSlime/Slimefun/Objects/SlimefunItem/items/SoulboundBackpack.java
@@ -1,10 +1,11 @@
-package me.mrCookieSlime.Slimefun.Objects.SlimefunItem;
+package me.mrCookieSlime.Slimefun.Objects.SlimefunItem.items;
 
 import org.bukkit.inventory.ItemStack;
 
+import io.github.thebusybiscuit.slimefun4.core.attributes.Soulbound;
 import me.mrCookieSlime.Slimefun.Lists.RecipeType;
 import me.mrCookieSlime.Slimefun.Objects.Category;
-import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.interfaces.Soulbound;
+import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.SlimefunBackpack;
 import me.mrCookieSlime.Slimefun.api.SlimefunItemStack;
 
 public class SoulboundBackpack extends SlimefunBackpack implements Soulbound {
